// apps/repoone/src/types/enums/AD_POSITIONS.ts
export enum AD_POSITIONS {
  CALC_TOP_MOB = 'CALC_TOP_MOB',
  CALC_TOP_SMALL = 'CALC_TOP_SMALL',
  // CALC_TOP_LARGE = 'CALC_TOP_LARGE',
  CALC_RIGHT_SKYSCRAPER = 'CALC_RIGHT_SKYSCRAPER',
  CALC_LEFT_SKYSCRAPER = 'CALC_LEFT_SKYSCRAPER',
  CALC_MOBILE_BOTTOM_SQUARE = 'CALC_MOBILE_BOTTOM_SQUARE',
  DESKTOP_BOTTOM = 'DESKTOP_BOTTOM',
  CALC_TOP_BANNER = 'CALC_TOP_BANNER',
  MOBILE_STICKY = 'MOBILE_STICKY',
  CONTENT_SIDE = 'CONTENT_SIDE',
}